// components/common/inputSearch/inputSearchStyle.scss.d.ts
export type Styles = {
  'buttonReset': string;
  'buttonSearch': string;
  'componentContainer': string;
  'iconCancel': string;
  'iconSearch': string;
  'inputSearch': string;
  'readOnlySearchValue': string;
  'rightButton': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;